CLN: Make bigquery backend not depend on impala backend
